service

+++ skinny-cat-8792/src/main/java/com/masai/service/OrderServiceimpl.java
@@ -1,7 +1,9 @@
 package com.masai.service;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import javax.persistence.criteria.Order;
@@ -9,6 +11,8 @@ import javax.persistence.criteria.Order;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.masai.dto.AdressDto;
+import com.masai.dto.ProductDto;
 import com.masai.exception.AddressException;
 import com.masai.exception.CartException;
 import com.masai.exception.LoginException;
@@ -16,9 +20,12 @@ import com.masai.exception.OrderExcetion;
 import com.masai.model.Customer;
 import com.masai.model.Orders;
 import com.masai.model.Session;
+import com.masai.repository.CartDao;
 import com.masai.repository.CustomerDao;
 import com.masai.repository.Orderdao;
 import com.masai.repository.SessionDao;
+import com.masai.model.Address;
+import com.masai.model.Cart;
 
 @Service
 public class OrderServiceimpl  implements OrderService{
@@ -33,6 +40,9 @@ public class OrderServiceimpl  implements OrderService{
 	 @Autowired
 	 private CustomerDao cus;
 	
+	 
+	 @Autowired
+	 private CartDao cartdao;
 	
 	@Override
 	public Orders addOrder(Orders order, String key) throws LoginException, CartException, OrderExcetion {
@@ -49,8 +59,65 @@ public class OrderServiceimpl  implements OrderService{
 		    }
 		    else
 		    {
-		    	return  od.save(order);
+		    	
+		    	
+		    	 Integer customerId =  ke.getUserId();
+		    	 
+		    	 Optional<Customer>  s = cus.findById(customerId);
+		    	 
+		    	 Address ss = s.get().getAddress();
+		    	 
+		    	 Orders curr  = new Orders();
+		    	 
+		    	 curr.setOrderDate(LocalDate.now());
+                 curr.setAddress( new AdressDto(ss.getStreetNo(), ss.getBulidingName(), ss.getCity(), ss.getState(), ss.getCountry(), ss.getPincode()));
+                 curr.setCustomer(s.get());
+                 curr.setOrderStatus("Order Confrimed");
+                 
+                 
+                 List<ProductDto> list = cartdao.findByCustomer(s.get()).getProducts();
+                 
+                 if( list.size() < 1) {
+    				 throw new CartException("Add product to the cart first...");
+    			 }
+                 
+                 
+                 
+                 List<ProductDto> product = new ArrayList<ProductDto>();
+                 
+                 Double total = 0.0;
+//                 for(ProductDto s: )
+		    	
+                 
+                 for(ProductDto proDto : list) {
+    				 
+    				 product.add(proDto);
+    				 
+    				 total += (proDto.getPrice() * proDto.getQuantity()) ;
+    				 
+    			 }
+                 
+                 curr.setTotal(total);	
+    			 curr.setPoder(product);
+    			 
+    			 
+    			 
+               Integer customerI =  ke.getUserId();
+		    	 
+		    	 Optional<Customer>  t = cus.findById(customerI);
+    			 
+    			 
+    			 Cart customerCart = cartdao.findByCustomer(t.get()) ;
+    			 
+    			 customerCart.setProducts(new ArrayList<>());
+    			 
+    			 cartdao.save(customerCart);
+    			 
+    			 return od.save(curr);
+                 
+                 
 		    }
+		    
 	
 	}
 
@@ -67,7 +134,16 @@ public class OrderServiceimpl  implements OrderService{
 	  }
 	  else
 	  {
-		  return od.save(order);
+		  Optional<Orders> opt=  od.findById(order.getOrderId());
+		  
+		  
+		  if(opt.isPresent()) {
+				return od.save(order);
+			}
+			else {
+				throw new OrderExcetion("Order does not exist");
+			}
+		  
 	  }
 		
 		
@@ -87,13 +163,18 @@ public class OrderServiceimpl  implements OrderService{
 			{
 			   Optional<Orders> ss = od.findById(id);
 			   
-			    Orders k = ss.get();
-			    
-			    od.deleteById(id);
-               
-			    return k;
-			    		
-			   
+			    if(ss.isEmpty())
+			    {
+			    	throw new  OrderExcetion("You Don't have orders prodcut");
+			    }
+			    else
+			    {
+			    	Orders se = ss.get();
+			    	od.deleteById(id);
+			    	
+			    	return se;
+			    }
+			     
 			}
 	}
